feat: Migrated Work Module to KMP
