+++ sync/work/src/androidProd/AndroidManifest.xml
