+++ sync/work/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/sync/status/StubSyncManager.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.sync.status
+
+import co.touchlab.kermit.Logger
+import com.google.samples.apps.nowinandroid.core.data.util.SyncManager
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flowOf
+
+private const val TAG = "StubSyncManager"
+
+class StubSyncManager: SyncManager {
+    override val isSyncing: Flow<Boolean>
+        get() = flowOf(false)
+
+    override fun requestSync() {
+        Logger.d("Requesting to sync", null, TAG)
+    }
+}
