--- sync/work/src/main/kotlin/com/google/samples/apps/nowinandroid/sync/status/StubSyncSubscriber.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright 2023 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.samples.apps.nowinandroid.sync.status
-
-import android.util.Log
-
-private const val TAG = "StubSyncSubscriber"
-
-/**
- * Stub implementation of [SyncSubscriber]
- */
-class StubSyncSubscriber : SyncSubscriber {
-    override suspend fun subscribe() {
-        Log.d(TAG, "Subscribing to sync")
-    }
-}
